requirements.txt
arrow==1.2.3
asgiref==3.5.2
async-timeout==4.0.2
Authlib==1.0.1
autopep8==1.6.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
commonmark==0.9.1
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.4
django-currentuser==0.5.3
django-filter==2.4.0
django-guardian==2.4.0
django-redis==5.2.0
djangorestframework==3.12.4
google-auth==2.10.0
idna==3.3
kubernetes==24.2.0
multi-key-dict==2.0.3
oauthlib==3.2.0
packaging==21.3
pbr==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-gitlab==3.8.0
python-jenkins==1.7.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==12.5.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.11
websocket-client==1.3.3
wrapt==1.14.1
